moved driving from periodic into a default command
driving in periodic was the issue that was causing auton file playback to be much worse, because the scheduler was sending 0 values to the motors every 20 ms in between the real recorded values, making ramping impossible and lagging the bot substantially. also changed it so that we record motor values now instead of joystick values (not sure if it makes much of a difference).